tests/wpt/tests/html/semantics/embedded-content/the-canvas-element/type.extend.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: type.extend</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>type.extend</h1>
<p class="desc">HTMLCanvasElement methods can be added, and the new methods used by canvases</p>

<p class="notes">Defined in "Web IDL" (draft)
<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("HTMLCanvasElement methods can be added, and the new methods used by canvases");
_addTest(function(canvas, ctx) {

window.HTMLCanvasElement.prototype.getZero = function () { return 0; };
_assertSame(canvas.getZero(), 0, "canvas.getZero()", "0");


});
</script>
